apps/agent/requirements-simple.txt
# Simple Agent Requirements (No LangChain)
celery==5.3.4
redis==5.0.1
openai==1.3.7
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
httpx==0.25.2
python-dotenv==1.0.0
pydantic==2.5.0
